terrabia-backend/requirements.txt
asgiref==3.11.0
async-timeout==5.0.1
attrs==25.4.0
autobahn==24.4.2
Automat==25.4.16
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
Django==5.2.8
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
inflection==0.5.1
msgpack==1.1.2
mysqlclient==2.2.7
packaging==25.0
pillow==12.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
service-identity==24.2.0
sqlparse==0.5.3
tomli==2.3.0
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
uritemplate==4.2.0
zope.interface==8.1.1
